requirements.txt
Django==5.0.6
djangorestframework==3.15.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
pillow==10.0.1
requests==2.31.0
pytz==2023.3
mysqlclient==2.2.0
gunicorn==21.2.0
flower==2.0.1
django-prometheus==2.3.1
django-celery-beat